move trade return value to types file
